Define a fish function to run commands in containers
A perennial problem for me is that I need to work inside containers I don't
really have much control over. For instance, I need to work with a build
image, based on an ancient CentOS (8.4!). As such, I can't do the "unidev"
thing and install tools and work _inside_ the image. Instead, I need to
stay _outside_ where my tools are, but then I can't run the compiled
binaries, because they are meant to run _inside_. This command bridges the
gap, giving me a way to map a command over the container, quite like an IO
functor: where the injected function runs _within_ the container, but it's
input and output are connected to the outside world.