Add support for automatic PATH munging
The current bp--run program has a flaw where the contents of the
build-debug/bin directory are not included. However, since I want to run
the command with a `nice` wrapper, I can no longer just put an environment
assignment in front of it, because `nice` doesn't know what to do with it.
Instead, in this commit, I add a little logic to automatically add a
certain directory to the PATH before running the command. This is
controlled by option "path", and is set to "build-debug/bin" by default.

NOTE: the directory is _relative_ to the _project_, not the current working
directory!