Clean up ll and l commands
In this commit, I clean up the ll command, so that all shells try lsd,
then exa, then ls to list long. I also try to keep the results as useful
as possible in all of those cases.

The new l command is meant to be a universal "look" tool on paths:
automatically dealing with them the right way. For now, it just deals
with directories vs regular files. Eventually, I want to teach it to
properly open URLs in the browser, images in a viewer, PDFs in their
viewer, and so on.