Fix weave
In this commit, I slightly alter the weave rule so that it won't accidentially trigger typst into closing the source block too early. The essential
problem is, if I print ``` in AWK to weave the same rule that defines weave, it will mistakenly close the block too early. However, if I break up the
string with a couple of extra quotes, AWK won't mind and typst won't see it as a close or open statement for a source block. This fixes the trouble.