feat(git-foreach): Clean up source and output
In this commit, I clean up the output by using SGR to color, bold, and
underline the directory names to make them more visually striking.
Further, I also ensure the output of the recursive commands is in TAB
column 2, both for more visual separation, and to make it easier to
parse the output if we want. E.g., just ignoring the output is as easy
as piping to `grep -v '\t'` or something similar.

I also clean up the source, using a clever GNU parallel trick of using
shell function exports to expose multiline functions to parallel. The
result is much cleaner and easier to understand, lacking all of the
silly escapes from before.