feat(log2tsv): Add ability to find log lines that are quoted
In this version, I add the ability to also parse log lines that use '\t'
instead of an actual TAB character to separate components. This allows
the program to find logs even if the original is quoted, where before it
would be impossible. I'm not sure if this feature needs to be
disableable or not yet, so I'm just hard-coding it for now.