feat(bt): Dump stack trace on a live program
Sometimes I need to see what a program is doing. The bt command uses GDB
to dump all threads' stack traaces, permitting me to sample the stack of
a program at any time. If I then do this a number of times, and group,
count, and sort, I'll have a nice poor man's profiler that produces
output in TSV format.