Add home-manager config to dotfiles
home-manager is a NIX-based declarative system for managing programs and program configuration per-user. I intend to use this to essentially replace most of the current dotfiles repo, assuming I can reliably install deterministic-nix or plain nix on a target. For example, my new work computer may not want to let me do that, in which case, I have to fall back on stow and the old system.