Add search comments and reorganize
In this commit, I add a table of contents that uses a simple searchable
tag for each item. I also make sure the sections stand out a bit more. I
anticipate this file to grow quite large as I add configurations to it,
so I want a better way to navigate it in my editor.