Create leap labels for home-manager config
In this commit, I add a pile of "leap labels", which are a low-effort
way to provide links within the document. The basic idea is to use a
commented string that starts with [, a number (I just use a line
number), ], and a description. The idea is that the numbers are GUIDs
within the file, and can be searched. Then, I place three copies of that
line: one at the TOC, one at the beginning of the section, and one at
the end. That way, I can use vim's '*' and '#' commands to skip around
pretty easily. I'm sure there's a better way (TM), but this seems to be
good enough for now.

Oops, I forgot to put this in the right commit.