Fix several issues with gpg-agent
One infuriating problem with default gpg is the fact that it wants to
use the GUI to prompt for pins. That is a problem when I'm logged in via
MOSH or SSH and can't _access_ the gui. This makes the pinentry program
use curses, so I can see it in my terminals from now on.